move around function withSessionConfig.
