+++ sql/core/src/main/scala/org/apache/spark/sql/DataFrameReader.scala
@@ -20,7 +20,6 @@ package org.apache.spark.sql
 import java.util.{Locale, Properties}
 
 import scala.collection.JavaConverters._
-import scala.collection.immutable
 
 import org.apache.spark.Partition
 import org.apache.spark.annotation.InterfaceStability
@@ -33,8 +32,8 @@ import org.apache.spark.sql.execution.datasources.{DataSource, FailureSafeParser
 import org.apache.spark.sql.execution.datasources.csv._
 import org.apache.spark.sql.execution.datasources.jdbc._
 import org.apache.spark.sql.execution.datasources.json.TextInputJsonDataSource
+import org.apache.spark.sql.execution.datasources.v2.DataSourceV2ConfigSupport
 import org.apache.spark.sql.execution.datasources.v2.DataSourceV2Relation
-import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.sources.v2._
 import org.apache.spark.sql.types.{StringType, StructType}
 import org.apache.spark.unsafe.types.UTF8String
@@ -171,7 +170,7 @@ class DataFrameReader private[sql](sparkSession: SparkSession) extends Logging {
     option("path", path).load(Seq.empty: _*) // force invocation of `load(...varargs...)`
   }
 
-  import DataFrameReader._
+  import DataSourceV2ConfigSupport._
 
   /**
    * Loads input in as a `DataFrame`, for data sources that support multiple paths.
@@ -742,26 +741,4 @@ class DataFrameReader private[sql](sparkSession: SparkSession) extends Logging {
 
   private val extraOptions = new scala.collection.mutable.HashMap[String, String]
 
-}
-
-private[sql] object DataFrameReader {
-
-  /**
-   * Helper method to filter session configs with config key that matches at least one of the given
-   * prefixes.
-   *
-   * @param cs the config key-prefixes that should be filtered.
-   * @param conf the session conf
-   * @return an immutable map that contains all the session configs that should be propagated to
-   *         the data source.
-   */
-  def withSessionConfig(
-      cs: ConfigSupport,
-      conf: SQLConf): immutable.Map[String, String] = {
-    val prefixes = cs.getConfigPrefixes
-    require(prefixes != null, "The config key-prefixes cann't be null.")
-    conf.getAllConfs.filterKeys { confKey =>
-      prefixes.asScala.exists(confKey.startsWith(_))
-    }
-  }
 }
